fix: fixing delete category

+++ Back-End/palmbooking/src/main/java/com/backend/palmbooking/Controller/CategoryController.java
@@ -3,6 +3,9 @@ package com.backend.palmbooking.Controller;
 import com.backend.palmbooking.Model.Category;
 import com.backend.palmbooking.Service.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -41,8 +44,14 @@ public class CategoryController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteCategoryByID(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteCategoryByID(@PathVariable Long id) {
+        Optional<Category> buscarCategoria = categoryService.getCategoryByID(id);
+
+        if (buscarCategoria.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         categoryService.deleteCategoryByID(id);
+        return ResponseEntity.status(204).build();
     }
